add quickcheck test and fix free list
This commit adds a quickcheck test designed to detect cases in which the
buddy allocator returns a region overlapping a previous allocation. That
behavior was not observed; however, the test did detect a bug in
free_list_find_remove(), in which the target block was found in the free
list but the return value indicated that it had not been. This bug has
been fixed.